backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
pymupdf>=1.23.0
openai>=1.12.0
python-dotenv>=1.0.0
supabase>=2.0.0
postgrest>=0.16.0
